Conflicts:
	pet/src/main/java/mutli/com/pet/erp/MemberServiceImpl.java
	pet/src/main/java/mutli/com/pet/erp/member.xml
	pet/src/main/webapp/WEB-INF/views/login.jsp
	pet/target/classes/mutli/com/pet/erp/MemberServiceImpl.class
	pet/target/classes/mutli/com/pet/erp/member.xml

+++ pet/src/main/java/mutli/com/pet/erp/MemberService.java
@@ -40,6 +40,6 @@ public interface MemberService {
 	List<SitterDTO> sitterList();
 	
 	// sitter rate
-	int sitter_rate_update(String sitter_id, String review_rate);
+	double sitter_rate_update(String sitter_id, String review_rate);
 	
 }
